move flow/rivertypes package to river/rivertypes
With the intent to add a `nonsensitive` function into the standard
library, rivertypes needs to be moved under river/ to allow it to
eventually remove import cycles.

Plus, given a `nonsensitive` function, the river/ package would be a
better home for the rivertypes package, since the River stdlib should
not import anything outside of River.
