+++ component/prometheus/exporter/postgres/postgres.go
@@ -3,9 +3,9 @@ package postgres
 import (
 	"github.com/grafana/agent/component"
 	"github.com/grafana/agent/component/prometheus/exporter"
-	"github.com/grafana/agent/pkg/flow/rivertypes"
 	"github.com/grafana/agent/pkg/integrations"
 	"github.com/grafana/agent/pkg/integrations/postgres_exporter"
+	"github.com/grafana/agent/pkg/river/rivertypes"
 	config_util "github.com/prometheus/common/config"
 )
