+++ component/remote/http/http_test.go
@@ -12,8 +12,8 @@ import (
 	"github.com/go-kit/log/level"
 	http_component "github.com/grafana/agent/component/remote/http"
 	"github.com/grafana/agent/pkg/flow/componenttest"
-	"github.com/grafana/agent/pkg/flow/rivertypes"
 	"github.com/grafana/agent/pkg/river"
+	"github.com/grafana/agent/pkg/river/rivertypes"
 	"github.com/grafana/agent/pkg/util"
 	"github.com/grafana/dskit/backoff"
 	"github.com/stretchr/testify/require"
