+++ pkg/river/rivertypes/optional_secret.go
